refactor(auth): move auth adapters to services/auth (phase 6F)
Move oidc_client, refresh, session_tokens, state_cookie, and deps from
components/auth/ to services/auth/. Old files become re-export shims.
Middleware stays in components/auth/ (phase 10). Tests updated to patch
the canonical module for stampede-guard tests.
