+++ openrag/components/auth/test_state_cookie.py
@@ -3,7 +3,7 @@
 import time
 
 import pytest
-from components.auth.state_cookie import StateCookiePayload, StateCookieSerializer
+from services.auth.state_cookie import StateCookiePayload, StateCookieSerializer
 
 SECRET = "test-secret-key-for-state-cookie"
